Add files via upload
X Education is facing a challenge with its lead conversion rate, which is currently at 30%. In order to improve this rate and achieve the CEO's target of an 80% conversion rate, the company requires us to develop a lead scoring model. The goal of this model is to assign a score to each lead that reflects their likelihood of converting to a customer, with higher scores indicating a higher chance of conversion. By prioritizing leads with higher scores, the company can focus its resources on the most promising leads and increase its overall conversion rate